move moe tensor
